OAK-11617: Provide oak-run commands to analyze and fix inconsistencies in the namespace registry
Moved unit test from oak-it to oak-core.
